users/urls.py
from django.urls import include, path
from rest_framework import routers
from . import views
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
    TokenVerifyView,
) 


router = routers.DefaultRouter()
router.register(r'users', views.UserViewSet)
router.register(r'groups', views.GroupViewSet)

# Wire up our API using automatic URL routing.
# Additionally, we include login URLs for the browsable API.
urlpatterns = [
    path('', include(router.urls)),
    path('api-auth/', include('rest_framework.urls', namespace='rest_framework')),
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/token/verify/', TokenVerifyView.as_view(), name='token_verify'),

    path('login/', views.LoginViev.as_view(), name='customer-login'),
    path('api/users/', views.UserCreate.as_view(), name='account-create'),
    path('api/users/user/create/profile/', views.UserCreateProfileView.as_view(), name='create-profile'),
    path('api/users/profile/detail/<int:pk>/', views.ProfileDetail.as_view(), name='account-detail'),
]